refactor: extract logic to ask user to select something into a function
